Updating spec to use decorator
This commit builds on the rebased commit of
e9697e4641ac50068fbda72e0bed94fc8adbdba4

The rebased commit was from an aging PR (#3881) that addressed #3775.
In the time between the submission of the original PR and the
revisitation, we updated some underlying aspects of Hyrax.

This is a quick fix to get a build working.  I'll rely on reviewers to
determine if its appropriate.

A question remains, should our `Hyrax::SolrDocument::OrderedMembers`
"be_a" SolrDocument?