Favoring helper method over presenter method
In discussions with @no_reply, I'm looking to trim the size of
presenters.  As envisioned, presenters should be more about gathering up
data to present.  Helpers shine when used for conditional rendering.

Since we don't ask presenters to "render themself" this inflection makes
sense; Let the helpers decide render logic.

This is an interstitial commit.  It preserves the backwards behavior.